Splitted up LogLabels into LogLabelStats and LogLabel
